backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
SQLAlchemy>=2.0
psycopg2-binary
pymysql
cryptography
passlib[bcrypt]
python-jose[cryptography]
pydantic-settings
email-validator
